+++ packages/nns/src/types/governance.options.ts
@@ -1,5 +1,5 @@
+import type { CanisterOptions } from "@dfinity/utils";
 import type { _SERVICE as GovernanceService } from "../../candid/governance";
-import type { CanisterOptions } from "./canister.options";
 
 export interface GovernanceCanisterOptions
   extends CanisterOptions<GovernanceService> {