+++ packages/sns/src/ledger.canister.ts
@@ -1,3 +1,4 @@
+import { createServices } from "@dfinity/utils";
 import type { _SERVICE as SnsLedgerService } from "../candid/icrc1_ledger";
 import { idlFactory as certifiedIdlFactory } from "../candid/icrc1_ledger.certified.idl";
 import { idlFactory } from "../candid/icrc1_ledger.idl";
@@ -5,7 +6,6 @@ import { Canister } from "./services/canister";
 import type { SnsCanisterOptions } from "./types/canister.options";
 import type { SnsTokenMetadataResponse } from "./types/ledger.responses";
 import type { QueryParams } from "./types/query.params";
-import { createServices } from "./utils/actor.utils";
 
 export class SnsLedgerCanister extends Canister<SnsLedgerService> {
   static create(options: SnsCanisterOptions<SnsLedgerService>) {