+++ packages/sns/src/swap.canister.ts
@@ -1,3 +1,4 @@
+import { createServices, fromNullable } from "@dfinity/utils";
 import type {
   BuyerState,
   GetBuyerStateRequest,
@@ -10,8 +11,6 @@ import { idlFactory } from "../candid/sns_swap.idl";
 import { Canister } from "./services/canister";
 import type { SnsCanisterOptions } from "./types/canister.options";
 import type { QueryParams } from "./types/query.params";
-import { createServices } from "./utils/actor.utils";
-import { fromNullable } from "./utils/did.utils";
 
 export class SnsSwapCanister extends Canister<SnsSwapService> {
   static create(options: SnsCanisterOptions<SnsSwapService>) {