+++ packages/sns/src/types/canister.options.ts
@@ -1,20 +1,8 @@
-import type { ActorSubclass, Agent } from "@dfinity/agent";
 import type { Principal } from "@dfinity/principal";
+import type { CanisterOptions } from "@dfinity/utils";
 
-// Note: only difference with nns canister options is canisterId is mandatory for Sns
-
-export interface CanisterOptions<T> {
-  // The agent to use when communicating with the canister.
-  agent?: Agent;
-  // The service to use when calling into the IC. Primarily overridden
-  // in test for mocking.
-  serviceOverride?: ActorSubclass<T>;
-  // The service to use when calling into the IC. Primarily overridden
-  // in test for mocking.
-  certifiedServiceOverride?: ActorSubclass<T>;
-}
-
-export interface SnsCanisterOptions<T> extends CanisterOptions<T> {
-  // The canister's ID.
+export interface SnsCanisterOptions<T>
+  extends Omit<CanisterOptions<T>, "canisterId"> {
+  // The canister's ID is mandatory in Sns canister classes.
   canisterId: Principal;
 }