+++ packages/utils/src/index.ts
@@ -1,3 +1,5 @@
+export type { CanisterOptions } from "./types/canister.options";
+export * from "./utils/actor.utils";
 export * from "./utils/agent.utils";
 export * from "./utils/asserts.utils";
 export * from "./utils/did.utils";